requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
# Optional for PostgreSQL / SQL Server:
# psycopg2-binary>=2.9.0
# pyodbc>=5.0.0
